Refactor PaginationHeaderComponent * move TOTAL_PAGES into /src/config/constants * improve code consistency * separate PaginationHeaderComponent.tsx sub-components to individual files
